Move get_contents() and get_sha1sum() to util.py
Both are general-use functions so moving to util.py.
